backend/requirements.txt
fastapi
uvicorn
torch
pandas
numpy
scikit-learn
datasets
google-generativeai
python-dotenv
sqlalchemy
pydantic
pickle5; python_version < "3.8"